requirements.txt
torch>=2.0.0
transformers>=4.52.4
datasets>=2.14.0
peft>=0.7.0
langgraph>=0.0.30
langchain>=0.1.0
numpy>=1.23.0
huggingface_hub=0.33.0
python-dotenv=1.1.0
